Split test_db.py into per-table test files
test_db.py had grown to 1411 lines covering nine unrelated concerns
(organizations, sources, jobs, records, record errors, datasets,
locations, faceted queries, slug protection) in one file. The
datagov_data_access library already solved this exact problem for its
own copy of these tests, splitting along the same lines into
tests/integration/db/test_{organization,harvest_source,harvest_job,
harvest_record,harvest_record_error,dataset,location,queries,
dataset_slug}.py. Mirror that layout here.

All 52 tests move over unchanged (verified 1:1 by function name and
by a full pytest run); only the two dataset-fixture helper functions
(_create_record_for_dataset, _dataset_payload) move from private
methods on the old TestDatabase class into module-level
create_record_for_dataset/dataset_payload functions in
tests/conftest.py, shared by test_dataset.py, again matching the
library's own tests/conftest.py.
