--- tests/integration/database/test_db.py
@@ -1,1411 +1,0 @@
-from datetime import datetime, timezone
-from unittest.mock import MagicMock, patch
-
-from freezegun import freeze_time
-from sqlalchemy import text
-
-from database.interface import PAGINATE_ENTRIES_PER_PAGE
-from database.models import (
-    DatasetViewCount,
-    HarvestJobError,
-    HarvestRecordError,
-    Locations,
-)
-
-
-@freeze_time("Jan 14th, 2012")
-class TestDatabase:
-    def _create_record_for_dataset(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-        identifier="dataset-popularity-record",
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        job_data_dcatus["harvest_source_id"] = source_data_dcatus["id"]
-        interface.add_harvest_job(job_data_dcatus)
-        return interface.add_harvest_record(
-            {
-                "identifier": identifier,
-                "harvest_job_id": job_data_dcatus["id"],
-                "harvest_source_id": source_data_dcatus["id"],
-                "status": "success",
-                "action": "create",
-                "source_raw": "{}",
-            }
-        )
-
-    def _dataset_payload(
-        self,
-        slug,
-        record,
-        organization_data,
-        source_data_dcatus,
-        translated_spatial=None,
-    ):
-        payload = {
-            "slug": slug,
-            "dcat": {"title": slug},
-            "organization_id": organization_data["id"],
-            "harvest_source_id": source_data_dcatus["id"],
-            "harvest_record_id": record.id,
-            "last_harvested_date": datetime.now(timezone.utc),
-        }
-
-        if translated_spatial is not None:
-            payload["translated_spatial"] = translated_spatial
-
-        return payload
-
-    def test_add_organization(self, interface, organization_data):
-        org = interface.add_organization(organization_data)
-
-        assert org is not None
-        assert org.name == "Test Org"
-
-    def test_get_all_organizations(self, interface, organization_data):
-        interface.add_organization(organization_data)
-
-        orgs = interface.get_all_organizations()
-        assert len(orgs) > 0
-        assert orgs[0].name == "Test Org"
-
-    def test_get_organization_by_slug(self, interface, organization_data):
-        interface.add_organization(organization_data)
-
-        org = interface.get_organization_by_slug(organization_data["slug"])
-        assert org is not None
-        assert org.id == organization_data["id"]
-
-    def test_get_organization_by_alias(self, interface, organization_data):
-        interface.add_organization(organization_data)
-
-        org = interface.get_organization_by_alias(organization_data["aliases"][0])
-        assert org is not None
-        assert org.id == organization_data["id"]
-
-    def test_get_organization_by_alias_not_found(self, interface, organization_data):
-        interface.add_organization(organization_data)
-
-        org = interface.get_organization_by_alias("nonexistent-alias")
-        assert org is None
-
-    def test_update_organization(self, interface, organization_data):
-        org = interface.add_organization(organization_data)
-
-        updates = {"name": "Updated Org"}
-        updated_org = interface.update_organization(org.id, updates)
-        assert updated_org.name == "Updated Org"
-
-    def test_delete_organization(self, interface, organization_data):
-        org = interface.add_organization(organization_data)
-
-        result = interface.delete_organization(org.id)
-        # ruff: noqa: E501
-        assert result == (
-            "Deleted organization with ID:d925f84d-955b-4cb7-812f-dcfd6681a18f successfully",
-            200,
-        )
-
-    def test_add_harvest_source(self, interface, organization_data, source_data_dcatus):
-        interface.add_organization(organization_data)
-        source = interface.add_harvest_source(source_data_dcatus)
-
-        assert source is not None
-        assert source.name == source_data_dcatus["name"]
-
-    def test_add_harvest_source_waf_collection_no_parent_url(
-        self, interface, organization_data, source_data_dcatus, caplog
-    ):
-        """A waf-collection source must have a non-null collection_parent_url."""
-        interface.add_organization(organization_data)
-        source_data_dcatus["source_type"] = "waf-collection"
-        source = interface.add_harvest_source(source_data_dcatus)
-        assert source is None
-        assert 'violates check constraint "wafcollectionparenturl"' in caplog.text
-
-    def test_add_harvest_source_waf_collection_with_parent_url(
-        self, interface, organization_data, source_data_dcatus
-    ):
-        """Can add a waf-collection source with parent URL."""
-        interface.add_organization(organization_data)
-        source_data_dcatus["source_type"] = "waf-collection"
-        source_data_dcatus["collection_parent_url"] = "fake-url"
-        source = interface.add_harvest_source(source_data_dcatus)
-        assert source is not None
-
-    def test_add_harvest_source_document_with_collection_parent_url(
-        self, interface, organization_data, source_data_dcatus, caplog
-    ):
-        """document harvest source cannot have collection_parent_url."""
-        interface.add_organization(organization_data)
-        source_data_dcatus["collection_parent_url"] = "fake-url"
-        source = interface.add_harvest_source(source_data_dcatus)
-        assert source is None
-        assert 'violates check constraint "wafcollectionparenturl"' in caplog.text
-
-    def test_get_all_harvest_sources(
-        self, interface, organization_data, source_data_dcatus
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-
-        sources = interface.get_all_harvest_sources()
-        assert len(sources) > 0
-        assert sources[0].name == source_data_dcatus["name"]
-
-    def test_get_harvest_source(self, interface, organization_data, source_data_dcatus):
-        interface.add_organization(organization_data)
-        source = interface.add_harvest_source(source_data_dcatus)
-
-        fetched_source = interface.get_harvest_source(source.id)
-        assert fetched_source is not None
-        assert fetched_source.name == source_data_dcatus["name"]
-
-    def test_update_harvest_source(
-        self, interface, organization_data, source_data_dcatus
-    ):
-        interface.add_organization(organization_data)
-        source = interface.add_harvest_source(source_data_dcatus)
-
-        updates = {
-            "name": "Updated Test Source",
-            "notification_emails": ["example@gmail.com", "another@yahoo.com"],
-        }
-
-        updated_source = interface.update_harvest_source(source.id, updates)
-        assert updated_source is not None
-        assert updated_source.name == updates["name"]
-        assert updated_source.notification_emails == [
-            "example@gmail.com",
-            "another@yahoo.com",
-        ]
-
-    def test_delete_harvest_source(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-        record_data_dcatus,
-    ):
-        # Add an organization
-        interface.add_organization(organization_data)
-
-        # Add a harvest source
-        source = interface.add_harvest_source(source_data_dcatus)
-        assert source is not None
-
-        # Case 1: Harvest source has no records, so it can be deleted successfully
-        response = interface.delete_harvest_source(source.id)
-        # ruff: noqa: E501
-        assert response == (
-            "Deleted harvest source with ID:2f2652de-91df-4c63-8b53-bfced20b276b successfully",
-            200,
-        )
-
-        # Refresh the session to avoid ObjectDeletedError
-        interface.db.expire_all()
-
-        deleted_source = interface.get_harvest_source(source.id)
-        assert deleted_source is None
-
-        # Case 2: Harvest source has records, so deletion should fail
-        # Add the harvest source again
-        source = interface.add_harvest_source(source_data_dcatus)
-        interface.add_harvest_job(job_data_dcatus)
-        record_data_dcatus[0]["ckan_id"] = "1234"
-        record = interface.add_harvest_record(record_data_dcatus[0])
-
-        response = interface.delete_harvest_source(source.id)
-        assert response == (
-            "Failed: 1 records in the Harvest source, please clear it first.",
-            409,
-        )
-
-        # Ensure the source still exists after failed deletion attempt
-        source_still_exists = interface.get_harvest_source(source.id)
-        assert source_still_exists is not None
-
-        # Case 3: Harvest source was cleared successfully which means
-        # all latest records are labelled as "delete" allowing harvest source
-        # deletion to occur
-        interface.update_harvest_record(record.id, {"action": "delete"})
-        response = interface.delete_harvest_source(source.id)
-        assert response == (
-            "Deleted harvest source with ID:2f2652de-91df-4c63-8b53-bfced20b276b successfully",
-            200,
-        )
-
-        deleted_source = interface.get_harvest_source(source.id)
-        assert deleted_source is None
-
-    def test_harvest_source_by_jobid(
-        self, interface, organization_data, source_data_dcatus, job_data_dcatus
-    ):
-        interface.add_organization(organization_data)
-        source = interface.add_harvest_source(source_data_dcatus)
-        job_data_dcatus["harvest_source_id"] = source.id
-
-        harvest_job = interface.add_harvest_job(job_data_dcatus)
-        harvest_source = interface.get_harvest_source_by_jobid(harvest_job.id)
-
-        assert source.id == harvest_source.id
-
-    def test_add_harvest_job_error(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-        job_error_data,
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        interface.add_harvest_job(job_data_dcatus)
-        harvest_job_error = interface.add_harvest_job_error(job_error_data)
-
-        assert isinstance(harvest_job_error, HarvestJobError)
-        assert harvest_job_error.message == job_error_data["message"]
-
-        db_harvest_job_error = interface.pget_harvest_job_errors(
-            filter=text(f"harvest_job_id = '{job_data_dcatus['id']}'")
-        )
-        assert db_harvest_job_error[0].type == job_error_data["type"]
-        assert db_harvest_job_error[0].id == harvest_job_error.id
-
-    def test_add_harvest_record_error(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-        record_data_dcatus,
-        record_error_data,
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        interface.add_harvest_job(job_data_dcatus)
-        interface.add_harvest_record(record_data_dcatus[2])
-
-        harvest_record_error = interface.add_harvest_record_error(record_error_data[0])
-        assert isinstance(harvest_record_error, HarvestRecordError)
-        assert harvest_record_error.message == record_error_data[0]["message"]
-
-        harvest_record_error_from_db = interface.get_harvest_error(
-            harvest_record_error.id
-        )
-        assert harvest_record_error.id == harvest_record_error_from_db.id
-        assert (
-            harvest_record_error.harvest_record_id
-            == harvest_record_error_from_db.harvest_record_id
-        )
-
-    def test_add_harvest_record(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-        record_data_dcatus,
-    ):
-        interface.add_organization(organization_data)
-        source = interface.add_harvest_source(source_data_dcatus)
-        harvest_job = interface.add_harvest_job(job_data_dcatus)
-
-        record = interface.add_harvest_record(record_data_dcatus[0])
-
-        assert record.harvest_source_id == source.id
-        assert record.harvest_job_id == harvest_job.id
-
-    def test_endpoint_pagnation(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-        record_data_dcatus,
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        interface.add_harvest_job(job_data_dcatus)
-
-        id_lookup_table = {}
-        for i in range(100):
-            record = record_data_dcatus[0].copy()
-            del record["id"]
-            record["identifier"] = f"test-identifier-{i}"
-            db_record = interface.add_harvest_record(record)
-            id_lookup_table[db_record.identifier] = db_record.id
-
-        # get first page
-        db_records = interface.pget_harvest_records(page=0)
-        assert len(db_records) == PAGINATE_ENTRIES_PER_PAGE
-        assert db_records[0].identifier == "test-identifier-0"
-        assert id_lookup_table[db_records[0].identifier] == db_records[0].id
-
-        # get second page
-        db_records = interface.pget_harvest_records(page=1)
-        assert len(db_records) == PAGINATE_ENTRIES_PER_PAGE
-        assert db_records[0].identifier == "test-identifier-10"
-        assert id_lookup_table[db_records[0].identifier] == db_records[0].id
-
-        # get first page again
-        db_records = interface.pget_harvest_records(page=0)
-        assert len(db_records) == PAGINATE_ENTRIES_PER_PAGE
-        assert db_records[0].identifier == "test-identifier-0"
-        assert id_lookup_table[db_records[0].identifier] == db_records[0].id
-
-        # don't paginate via feature flag
-        db_records = interface.pget_harvest_records(paginate=False)
-        assert len(db_records) == 100
-        assert id_lookup_table[db_records[50].identifier] == db_records[50].id
-
-        # get page 6 (r. 100 - 119), which is out of bounds / empty
-        db_records = interface.pget_harvest_records(page=11)
-        assert len(db_records) == 0
-
-        db_records = interface.pget_harvest_records(
-            facets=f"id eq {id_lookup_table['test-identifier-0']}"
-        )
-        assert len(db_records) == 1
-        assert db_records[0].harvest_job_id == job_data_dcatus["id"]
-
-    def test_endpoint_order_by(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-        record_data_dcatus,
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        interface.add_harvest_job(job_data_dcatus)
-
-        id_lookup_table = {}
-        for i in range(100):
-            record = record_data_dcatus[0].copy()
-            del record["id"]
-            record["identifier"] = f"test-identifier-{i}"
-            db_record = interface.add_harvest_record(record)
-            id_lookup_table[db_record.identifier] = db_record.id
-
-        # get results in ascending order, and confirm it is default
-        db_records = interface.pget_harvest_records()
-        assert db_records[0].identifier == "test-identifier-0"
-        assert id_lookup_table[db_records[0].identifier] == db_records[0].id
-
-        # get results in descending order
-        db_records = interface.pget_harvest_records(order_by="desc")
-        assert db_records[0].identifier == "test-identifier-99"
-        assert id_lookup_table[db_records[0].identifier] == db_records[0].id
-
-    def test_endpoint_count(
-        self, interface_with_fixture_json, job_data_dcatus, record_data_dcatus
-    ):
-        interface = interface_with_fixture_json
-        job_id = job_data_dcatus["id"]
-        count = interface.get_harvest_record_errors_by_job(
-            job_id,
-            count=True,
-        )
-        assert (
-            count
-            # two errors for each record
-            == 2
-            * len(
-                [record for record in record_data_dcatus if record["status"] == "error"]
-            )
-            == 16
-        )
-
-    def test_endpoint_count_for_non_paginated_methods(
-        self, interface_with_fixture_json, source_data_dcatus, record_data_dcatus
-    ):
-        interface = interface_with_fixture_json
-        count = interface.get_latest_harvest_records_by_source_orm(
-            source_data_dcatus["id"],
-            count=True,
-        )
-        assert (
-            count
-            == len(
-                [
-                    record
-                    for record in record_data_dcatus
-                    if record["status"] == "success"
-                ]
-            )
-            == 2
-        )
-
-    def test_sync_count_for_non_paginated_methods(
-        self, interface_with_fixture_json, source_data_dcatus, record_data_dcatus
-    ):
-        interface = interface_with_fixture_json
-
-        # test sync count by adding a valid record without a ckan_id
-        interface.add_harvest_record(
-            {
-                "identifier": "test_identifier-11",
-                "harvest_job_id": "6bce761c-7a39-41c1-ac73-94234c139c76",
-                "harvest_source_id": "2f2652de-91df-4c63-8b53-bfced20b276b",
-                "action": "create",
-                "status": "success",
-            }
-        )
-
-        count = interface.get_latest_harvest_records_by_source_orm(
-            source_data_dcatus["id"],
-            count=True,
-        )
-
-        assert (count - 1) == 2
-
-    def test_errors_by_job(
-        self,
-        interface_with_multiple_sources,
-        job_data_dcatus,
-        job_data_dcatus_2,
-        record_error_data,
-        record_error_data_2,
-    ):
-        interface = interface_with_multiple_sources
-        job_id = job_data_dcatus["id"]
-        count = interface.get_harvest_record_errors_by_job(
-            job_id,
-            count=True,
-        )
-        all_errors_count = interface.pget_harvest_record_errors(
-            count=True,
-        )
-        assert count == len(record_error_data)
-        assert all_errors_count == len(record_error_data) + len(record_error_data_2)
-
-    def test_record_errors_by_severity(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-        record_data_dcatus,
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        job = interface.add_harvest_job(job_data_dcatus)
-        record = interface.add_harvest_record(record_data_dcatus[2])
-
-        interface.add_harvest_record_error(
-            {
-                "message": "an error",
-                "type": "TestException",
-                "harvest_job_id": job.id,
-                "harvest_record_id": record.id,
-                "severity": "error",
-            }
-        )
-        interface.add_harvest_record_error(
-            {
-                "message": "a warning",
-                "type": "TestException",
-                "harvest_job_id": job.id,
-                "harvest_record_id": record.id,
-                "severity": "warning",
-            }
-        )
-
-        # by_job: default returns only errors, None returns all
-        errors = interface.get_harvest_record_errors_by_job(job.id, per_page=999)
-        assert {e[0].severity for e in errors} == {"error"}
-        all_issues = interface.get_harvest_record_errors_by_job(
-            job.id, severity=None, per_page=999
-        )
-        assert {e[0].severity for e in all_issues} == {"error", "warning"}
-        warnings = interface.get_harvest_record_errors_by_job(
-            job.id, severity="warning", per_page=999
-        )
-        assert {e[0].severity for e in warnings} == {"warning"}
-
-        # by_record: same defaulting behavior
-        assert {
-            e.severity for e in interface.get_harvest_record_errors_by_record(record.id)
-        } == {"error"}
-        assert {
-            e.severity
-            for e in interface.get_harvest_record_errors_by_record(
-                record.id, severity=None
-            )
-        } == {"error", "warning"}
-
-        # pget: default returns only errors, None returns all
-        assert {
-            e.severity for e in interface.pget_harvest_record_errors(paginate=False)
-        } == {"error"}
-        assert {
-            e.severity
-            for e in interface.pget_harvest_record_errors(severity=None, paginate=False)
-        } == {"error", "warning"}
-
-        # for_view: filters by severity (severity is not surfaced in the rows).
-        # default returns only the error row; None returns both.
-        view_errors = interface.get_harvest_record_errors_by_job(
-            job.id, for_view=True, per_page=999
-        )
-        assert len(view_errors) == 1
-        view_all = interface.get_harvest_record_errors_by_job(
-            job.id, severity=None, for_view=True, per_page=999
-        )
-        assert len(view_all) == 2
-
-    def test_get_harvest_record_issues(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-        record_data_dcatus,
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        job = interface.add_harvest_job(job_data_dcatus)
-        record = interface.add_harvest_record(record_data_dcatus[2])
-
-        interface.add_harvest_record_error(
-            {
-                "message": "an error",
-                "type": "TestException",
-                "harvest_job_id": job.id,
-                "harvest_record_id": record.id,
-                "severity": "error",
-            }
-        )
-        interface.add_harvest_record_error(
-            {
-                "message": "a warning",
-                "type": "TestException",
-                "harvest_job_id": job.id,
-                "harvest_record_id": record.id,
-                "severity": "warning",
-            }
-        )
-
-        # returns both errors and warnings
-        issues = interface.get_harvest_record_issues(job.id, per_page=999)
-        assert {i[0].severity for i in issues} == {"error", "warning"}
-        # honors the count kwarg
-        assert interface.get_harvest_record_issues(job.id, count=True) == 2
-
-    def test_record_errors_summary_by_job(
-        self,
-        interface_with_multiple_sources,
-        job_data_dcatus,
-        record_error_data,
-    ):
-        interface = interface_with_multiple_sources
-        job_id = job_data_dcatus["id"]
-        summary = interface.get_record_errors_summary_by_job(
-            job_id,
-        )
-        assert sum(count for count in summary.values()) == len(
-            interface.get_harvest_record_errors_by_job(job_id, per_page=999)
-        )
-
-    def test_add_harvest_job_with_id(
-        self, interface, organization_data, source_data_dcatus, job_data_dcatus
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        job = interface.add_harvest_job(job_data_dcatus)
-        assert job.id == job_data_dcatus["id"]
-        assert job.status == job_data_dcatus["status"]
-        assert job.harvest_source_id == job_data_dcatus["harvest_source_id"]
-
-    def test_add_harvest_job_without_id(
-        self, interface, organization_data, source_data_dcatus, job_data_dcatus
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-
-        job_data_dcatus_id = job_data_dcatus["id"]
-        del job_data_dcatus["id"]
-        job = interface.add_harvest_job(job_data_dcatus)
-        assert job.id
-        assert job.id != job_data_dcatus_id
-        assert job.status == job_data_dcatus["status"]
-        assert job.harvest_source_id == job_data_dcatus["harvest_source_id"]
-
-    def test_add_harvest_job_records_warned(
-        self, interface, organization_data, source_data_dcatus, job_data_dcatus
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        job_data_dcatus["records_warned"] = 3
-
-        job = interface.add_harvest_job(job_data_dcatus)
-        assert job.records_warned == 3
-
-    def test_update_harvest_job_records_warned(
-        self, interface_no_jobs, job_data_dcatus
-    ):
-        interface_no_jobs.add_harvest_job(job_data_dcatus)
-        updated = interface_no_jobs.update_harvest_job(
-            job_data_dcatus["id"], {"records_warned": 42}
-        )
-        assert updated.records_warned == 42
-
-    def test_get_all_harvest_jobs_by_facet(
-        self, source_data_dcatus, interface_with_multiple_jobs
-    ):
-        source_id = source_data_dcatus["id"]
-        filtered_list = interface_with_multiple_jobs.pget_harvest_jobs(
-            facets=f"status eq new, harvest_source_id eq {source_id}"
-        )
-        assert len(filtered_list) == 3
-        assert filtered_list[0].status == "new"
-        assert filtered_list[0].harvest_source_id == source_data_dcatus["id"]
-
-    def get_new_harvest_jobs_in_past(self, interface_with_multiple_jobs):
-        filtered_job_list = interface_with_multiple_jobs.get_new_harvest_jobs_in_past()
-        all_jobs_list = interface_with_multiple_jobs.get_all_harvest_jobs()
-        assert len(all_jobs_list) == 24
-        assert len(filtered_job_list) == 2
-        assert (
-            len(
-                [
-                    x
-                    for x in all_jobs_list
-                    if x["status"] == "new"
-                    and x["date_created"].replace(
-                        tzinfo=timezone.utc
-                    )  # TODO should we be pushing to UTC in db?
-                    < datetime.now(timezone.utc)
-                ]
-            )
-            == 2
-        )
-
-    def test_get_new_harvest_jobs_by_source_in_future(
-        self, interface_with_multiple_jobs
-    ):
-        all_jobs_list = interface_with_multiple_jobs.pget_harvest_jobs(paginate=False)
-        source_id = all_jobs_list[0].harvest_source_id
-        filtered_job_list = (
-            interface_with_multiple_jobs.get_new_harvest_jobs_by_source_in_future(
-                source_id
-            )
-        )
-        assert len(all_jobs_list) == 12
-        assert len(filtered_job_list) == 3
-        assert (
-            len(
-                [
-                    x
-                    for x in all_jobs_list
-                    if x.status == "new"
-                    and x.date_created.replace(
-                        tzinfo=timezone.utc
-                    )  # TODO should we be pushing to UTC in db?
-                    > datetime.now(timezone.utc)
-                    and x.harvest_source_id == source_id
-                ]
-            )
-            == 3
-        )
-
-    def test_filter_jobs_by_faceted_filter(
-        self, source_data_dcatus, interface_with_multiple_jobs
-    ):
-        faceted_list = interface_with_multiple_jobs.pget_harvest_jobs(
-            facets="status eq new"
-        )
-
-        assert len(faceted_list) == 3
-        assert len([x for x in faceted_list if x.status == "new"]) == 3
-        assert (
-            len(
-                [
-                    x
-                    for x in faceted_list
-                    if x.harvest_source_id == source_data_dcatus["id"]
-                ]
-            )
-            == 3
-        )
-
-    def test_delete_harvest_job(
-        self,
-        interface_no_jobs,
-        job_data_dcatus,
-    ):
-        interface_no_jobs.add_harvest_job(job_data_dcatus)
-        res = interface_no_jobs.delete_harvest_job(job_data_dcatus["id"])
-        assert isinstance(res, tuple)
-        assert res[0] == "Harvest job deleted successfully"
-        assert res[1] == 200
-
-    def test_get_latest_harvest_records(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        source_data_dcatus_2,
-        job_data_dcatus,
-        job_data_dcatus_2,
-        latest_records,
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        # another source for querying against. see last record in
-        # `latest_records` fixture
-        interface.add_harvest_source(source_data_dcatus_2)
-        interface.add_harvest_job(job_data_dcatus)
-        interface.add_harvest_job(job_data_dcatus_2)
-
-        for record in latest_records:
-            interface.add_harvest_record(record)
-
-        latest_records = interface.get_latest_harvest_records_by_source(
-            source_data_dcatus["id"]
-        )
-
-        # remove volatile fields so compare works
-        for record in latest_records:
-            del record["id"]
-            record.pop("dataset_slug", None)
-
-        expected_records = [
-            {
-                "identifier": "a",
-                "source_hash": None,
-                "date_created": datetime(2024, 3, 1, 0, 0, 0, 1000),
-                "date_finished": None,
-                "ckan_id": None,
-                "action": "update",
-            },
-            {
-                "identifier": "b",
-                "source_hash": None,
-                "date_created": datetime(2024, 3, 1, 0, 0, 0, 1000),
-                "date_finished": None,
-                "ckan_id": None,
-                "action": "create",
-            },
-            {
-                "identifier": "c",
-                "source_hash": None,
-                "date_created": datetime(2024, 5, 1, 0, 0, 0, 1000),
-                "date_finished": None,
-                "ckan_id": None,
-                "action": "create",
-            },
-            {
-                "identifier": "e",
-                "source_hash": None,
-                "date_created": datetime(2024, 4, 3, 0, 0, 0, 1000),
-                "date_finished": None,
-                "ckan_id": None,
-                "action": "create",
-            },
-        ]
-
-        assert len(latest_records) == 4
-        # make sure there aren't records that are different
-        assert not any(x != y for x, y in zip(latest_records, expected_records))
-
-    def test_delete_outdated_records(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        source_data_dcatus_2,
-        job_data_dcatus,
-        job_data_dcatus_2,
-        latest_records,
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        # another source for querying against. see last records in
-        # `latest_records` fixture
-        interface.add_harvest_source(source_data_dcatus_2)
-        interface.add_harvest_job(job_data_dcatus)
-        interface.add_harvest_job(job_data_dcatus_2)
-        records = [interface.add_harvest_record(record) for record in latest_records]
-
-        # only adding 1 record error for simplicity
-        # we have access to all record errors via relationship in HarvestRecord
-        error_data = {
-            "message": "record is invalid",
-            "type": "ValidationException",
-            "date_created": datetime.now(timezone.utc),
-            "harvest_record_id": records[-2].id,
-            "harvest_job_id": records[-2].harvest_job_id,
-        }
-        interface.add_harvest_record_error(error_data)
-
-        latest_records_from_db1 = interface.get_latest_harvest_records_by_source(
-            source_data_dcatus["id"]
-        )
-
-        latest_records_from_db2 = interface.get_latest_harvest_records_by_source(
-            source_data_dcatus_2["id"]
-        )
-
-        outdated_records = interface.get_all_outdated_records(90)
-        assert len(outdated_records) == 7
-
-        all_records = (
-            len(latest_records_from_db1)
-            + len(latest_records_from_db2)
-            + len(outdated_records)
-        )
-        # latest records for all harvest sources (2) and all outdated records
-        # should be equal to the original fixture count
-        assert all_records == len(latest_records)
-
-        # we want outdated records for ALL harvest sources. this is harvest source 2
-        hs2_outdated = next(r for r in outdated_records if r.identifier == "f")
-        assert len(hs2_outdated.errors) == 1
-        for record in outdated_records:
-            interface.delete_harvest_record(record_id=record.id)
-
-        # make sure only the outdated records and associated errors were deleted
-        db_records = interface.pget_harvest_records(count=True)
-        assert db_records == len(latest_records_from_db1) + len(latest_records_from_db2)
-
-        db_record_errors = interface.pget_harvest_record_errors(count=True)
-
-        # It should be expected the 1 error is still present
-        assert db_record_errors == 1
-
-    def test_deleting_job_deletes_errors(
-        self,
-        interface_with_fixture_json,
-        job_data_dcatus,
-        record_error_data,
-    ):
-        """
-        Test that confirms that HarvestRecordErrors are deleted on
-        associated HarvestJob deletion.
-        """
-        interface = interface_with_fixture_json
-        job_id = job_data_dcatus["id"]
-        # Confirm that errors are created
-        count = interface.get_harvest_record_errors_by_job(
-            job_id,
-            count=True,
-        )
-        assert count == len(record_error_data)
-
-        # Delete harvest job with errors
-        interface.delete_harvest_job(job_id)
-        count = interface.get_harvest_record_errors_by_job(
-            job_id,
-            count=True,
-        )
-        # Confirm that HarvestRecordErrors are deleted
-        # with the HarvestJob
-        assert count == 0
-
-    def test_harvest_record_error_remains(
-        self,
-        interface_with_fixture_json,
-        job_data_dcatus,
-        record_data_dcatus,
-        record_error_data,
-    ):
-        """
-        Test to confirm that HarvestRecordErrors are not deleted when
-        associated HarvestRecord is deleted.
-        """
-        interface = interface_with_fixture_json
-        job_id = job_data_dcatus["id"]
-        # Confirm that errors are created
-        error_count = interface.pget_harvest_record_errors(
-            count=True,
-        )
-        assert error_count == len(record_error_data)
-        harvest_job_records = interface.get_harvest_records_by_job(
-            job_id, paginate=False
-        )
-
-        # Confirm we have existing records and that they equal
-        # the number from our baseline
-        assert len(harvest_job_records) == len(record_data_dcatus)
-        # Delete HarvestRecords
-        for record in harvest_job_records:
-            interface.delete_harvest_record(record_id=record.id)
-
-        # Confirm no more HarvestRecords exist but we do have HarvestRecordErrors
-        harvest_job_records = interface.get_harvest_records_by_job(
-            job_id, paginate=False
-        )
-        error_count = interface.pget_harvest_record_errors(
-            count=True,
-        )
-        assert len(harvest_job_records) != len(record_data_dcatus)
-        assert error_count == len(record_error_data)
-
-    def test_faceted_builder_queries(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-        record_data_dcatus,
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        interface.add_harvest_job(job_data_dcatus)
-
-        id_lookup_table = {}
-        for i in range(100):
-            record = record_data_dcatus[0].copy()
-            del record["id"]
-            record["identifier"] = f"test-identifier-{i}"
-            db_record = interface.add_harvest_record(record)
-            id_lookup_table[db_record.identifier] = db_record.id
-
-        # source id, no facets
-        db_records = interface.get_harvest_records_by_source(source_data_dcatus["id"])
-        assert len(db_records) == PAGINATE_ENTRIES_PER_PAGE
-        assert db_records[0].identifier == "test-identifier-0"
-        assert id_lookup_table[db_records[0].identifier] == db_records[0].id
-
-        # source id, plus page kwarg
-        db_records = interface.get_harvest_records_by_source(
-            source_data_dcatus["id"],
-            page=1,
-        )
-        assert len(db_records) == PAGINATE_ENTRIES_PER_PAGE
-        assert db_records[0].identifier == "test-identifier-10"
-
-        # source id, plus pagination flag
-        db_records = interface.get_harvest_records_by_source(
-            source_data_dcatus["id"], paginate=False
-        )
-        assert len(db_records) == 100
-
-        # source id, plus kwargs to return only count
-        db_records = interface.get_harvest_records_by_source(
-            source_data_dcatus["id"],
-            count=True,
-        )
-        assert db_records == 100
-
-        # source id, plus extra filter_text facet
-        db_records = interface.get_harvest_records_by_source(
-            source_data_dcatus["id"],
-            facets=f"id eq {id_lookup_table['test-identifier-0']}",
-        )
-        assert len(db_records) == 1
-
-        # source id, plus extra filter_text facet, plus kwargs to return only count
-        db_records = interface.get_harvest_records_by_source(
-            source_data_dcatus["id"],
-            facets=f"id eq {id_lookup_table['test-identifier-0']}",
-            count=True,
-        )
-        assert db_records == 1
-
-        # source id, plus two facets
-        db_records = interface.get_harvest_records_by_source(
-            source_data_dcatus["id"],
-            facets=f"id eq {id_lookup_table['test-identifier-4']},identifier eq test-identifier-4",  # noqa E501
-            count=True,
-        )
-        assert db_records == 1
-
-    def test_get_geo_from_string(
-        self, interface, named_location_us, named_location_stoneham
-    ):
-        stoneham = Locations(
-            **{
-                "id": "34333",
-                "type": "us_postalcode",
-                "name": "2180",
-                "display_name": "Stoneham, MA (02180)",
-                "the_geom": "0103000020E61000000100000005000000BA6B09F9A0C751C046B6F3FDD4384540BA6B09F9A0C751C08E06F016484045401B9E5E29CBC451C08E06F016484045401B9E5E29CBC451C046B6F3FDD4384540BA6B09F9A0C751C046B6F3FDD4384540",  # noqa E501
-                "type_order": "4",
-            }
-        )
-        interface.db.add(stoneham)
-        interface.db.commit()
-
-        # Expect to find th US, loaded on test setup
-        geojson_str = interface.get_geo_from_string("United States")
-        assert geojson_str == named_location_us
-
-        # Expect to find Stoneham, loaded in this test
-        geojson_str = interface.get_geo_from_string("Stoneham")
-        assert geojson_str == named_location_stoneham
-
-        # Do not expect the following strings to match
-        assert interface.get_geo_from_string("not exists") is None
-        assert interface.get_geo_from_string("US, Virginia, Fairfax, Reston") is None
-
-    def test_extract_unsupported_schema(
-        self,
-        interface,
-        organization_data,
-        source_data_waf_csdgm,
-    ):
-        """attempts to add a 'csdgm' harvest source which is unsupported"""
-        interface.add_organization(organization_data)
-        # we can't add the source because we use an enum for schema type
-        assert interface.add_harvest_source(source_data_waf_csdgm) is None
-
-    def test_insert_dataset_view_counts(self, interface, view_count_datasets):
-        # add record to show delete
-        dataset_view_count = DatasetViewCount(dataset_slug="a", view_count=1)
-        interface.db.add(dataset_view_count)
-        interface.db.commit()
-
-        datasets = interface.db.query(DatasetViewCount).all()
-        assert len(datasets) == 1
-
-        interface.insert_view_counts_of_datasets(view_count_datasets)
-
-        datasets = interface.db.query(DatasetViewCount).all()
-        assert len(datasets) == 3  # b,c,d
-
-    def test_insert_dataset_sets_popularity_from_view_counts(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-    ):
-        record = self._create_record_for_dataset(
-            interface,
-            organization_data,
-            source_data_dcatus,
-            job_data_dcatus,
-            identifier="dataset-popularity-insert",
-        )
-        slug = "dataset-popularity-insert"
-        interface.db.add(DatasetViewCount(dataset_slug=slug, view_count=9))
-        interface.db.commit()
-
-        translated_geojson = {"type": "Point", "coordinates": [-77.0, 38.9]}
-
-        dataset = interface.insert_dataset(
-            self._dataset_payload(
-                slug,
-                record,
-                organization_data,
-                source_data_dcatus,
-                translated_spatial=translated_geojson,
-            )
-        )
-
-        assert dataset.popularity == 9
-        assert dataset.translated_spatial == translated_geojson
-
-    def test_upsert_dataset_updates_popularity_from_view_counts(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-    ):
-        record = self._create_record_for_dataset(
-            interface,
-            organization_data,
-            source_data_dcatus,
-            job_data_dcatus,
-            identifier="dataset-popularity-upsert",
-        )
-        slug = "dataset-popularity-upsert"
-        interface.db.add(DatasetViewCount(dataset_slug=slug, view_count=4))
-        interface.db.commit()
-
-        base_geojson = {"type": "Point", "coordinates": [-80.0, 25.0]}
-        base_payload = self._dataset_payload(
-            slug,
-            record,
-            organization_data,
-            source_data_dcatus,
-            translated_spatial=base_geojson,
-        )
-        interface.insert_dataset(base_payload)
-
-        # simulate analytics refresh increasing popularity
-        view_count = (
-            interface.db.query(DatasetViewCount).filter_by(dataset_slug=slug).first()
-        )
-        view_count.view_count = 12
-        interface.db.commit()
-
-        updated_geojson = {"type": "Point", "coordinates": [-81.0, 26.0]}
-        updated_payload = {
-            **base_payload,
-            "dcat": {"title": f"{slug}-updated"},
-            "last_harvested_date": datetime.now(timezone.utc),
-            "translated_spatial": updated_geojson,
-        }
-        dataset = interface.upsert_dataset(updated_payload)
-
-        assert dataset.popularity == 12
-        assert dataset.translated_spatial == updated_geojson
-
-    def test_count_missing_outdated_datasets(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        job_data_dcatus["harvest_source_id"] = source_data_dcatus["id"]
-        interface.add_harvest_job(job_data_dcatus)
-
-        # missing dataset harvest record
-        interface.add_harvest_record(
-            {
-                "identifier": "test",
-                "harvest_job_id": job_data_dcatus["id"],
-                "harvest_source_id": source_data_dcatus["id"],
-                "status": "success",
-                "action": "create",
-                "source_raw": "{}",
-            }
-        )
-
-        # dataset outdated harvest records
-        outdated_record = interface.add_harvest_record(
-            {
-                "identifier": "test-other",
-                "harvest_job_id": job_data_dcatus["id"],
-                "harvest_source_id": source_data_dcatus["id"],
-                "status": "success",
-                "action": "create",
-                "source_raw": "{}",
-                "date_created": datetime(2000, 3, 1, 0, 0, 0, 1000),
-            }
-        )
-
-        update_payload = {
-            "slug": "test-other",
-            "dcat": {"title": "Updated via harvest"},
-            "organization_id": organization_data["id"],
-            "harvest_source_id": source_data_dcatus["id"],
-            "harvest_record_id": outdated_record.id,
-            "last_harvested_date": datetime(2000, 3, 1, 0, 0, 0, 1000),
-        }
-
-        interface.upsert_dataset(update_payload)
-
-        interface.add_harvest_record(
-            {
-                "identifier": "test-other",
-                "harvest_job_id": job_data_dcatus["id"],
-                "harvest_source_id": source_data_dcatus["id"],
-                "status": "success",
-                "action": "update",
-                "source_raw": "{}",
-            }
-        )
-
-        # 1 missing and 1 outdated dataset
-        count, sample_records = (
-            interface.get_missing_or_outdated_dataset_count_and_sample()
-        )
-        assert count == 2 & len(sample_records) == 2
-
-
-@freeze_time("Mar 20th, 2026")
-class TestDatasetSlugProtection:
-    """
-    Tests ensuring the harvest cannot overwrite a dataset slug
-    that has been edited by a user via the dataset detail view.
-    """
-
-    def _setup_dataset(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-        slug,
-    ):
-        interface.add_organization(organization_data)
-        interface.add_harvest_source(source_data_dcatus)
-        job_data_dcatus["harvest_source_id"] = source_data_dcatus["id"]
-        interface.add_harvest_job(job_data_dcatus)
-        record = interface.add_harvest_record(
-            {
-                "identifier": slug,
-                "harvest_job_id": job_data_dcatus["id"],
-                "harvest_source_id": source_data_dcatus["id"],
-                "status": "success",
-                "action": "create",
-                "source_raw": "{}",
-            }
-        )
-        payload = {
-            "slug": slug,
-            "dcat": {"title": slug},
-            "organization_id": organization_data["id"],
-            "harvest_source_id": source_data_dcatus["id"],
-            "harvest_record_id": record.id,
-            "last_harvested_date": datetime.now(timezone.utc),
-        }
-        dataset = interface.insert_dataset(payload)
-        return dataset, record
-
-    def test_upsert_does_not_change_existing_slug(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-    ):
-        """
-        upsert_dataset must not alter the slug column of an existing row.
-
-        Even when a payload containing the same slug is passed in, the slug
-        must remain unchanged after the ON CONFLICT UPDATE path executes.
-        """
-        original_slug = "slug-protection-no-change"
-        dataset, record = self._setup_dataset(
-            interface,
-            organization_data,
-            source_data_dcatus,
-            job_data_dcatus,
-            original_slug,
-        )
-
-        # Re-submit with updated dcat slug in payload is the same
-        update_payload = {
-            "slug": original_slug,
-            "dcat": {"title": "Updated via harvest"},
-            "organization_id": organization_data["id"],
-            "harvest_source_id": source_data_dcatus["id"],
-            "harvest_record_id": record.id,
-            "last_harvested_date": datetime.now(timezone.utc),
-        }
-        result = interface.upsert_dataset(update_payload)
-
-        assert result.slug == original_slug
-        assert result.dcat == {"title": "Updated via harvest"}
-
-    def test_user_edited_slug_survives_harvest_upsert(
-        self,
-        interface,
-        organization_data,
-        source_data_dcatus,
-        job_data_dcatus,
-    ):
-        """
-        A user-edited slug must not be overwritten when harvest upserts.
-        """
-        harvest_slug = "slug-protection-harvest-generated"
-        dataset, record = self._setup_dataset(
-            interface,
-            organization_data,
-            source_data_dcatus,
-            job_data_dcatus,
-            harvest_slug,
-        )
-
-        # Simulate user editing the slug via the dataset detail form.
-        user_slug = "slug-protection-custom-user"
-        mock_client = MagicMock()
-        mock_client.index_datasets.return_value = (1, 0, [])
-        with patch(
-            "search.client.OpenSearchClient.from_environment",
-            return_value=mock_client,
-        ):
-            interface.update_dataset_slug(dataset.id, user_slug)
-
-        # On the next harvest run the harvester queries the DB for the
-        # current slug
-        reharvest_payload = {
-            "slug": user_slug,
-            "dcat": {"title": "Re-harvested Title"},
-            "organization_id": organization_data["id"],
-            "harvest_source_id": source_data_dcatus["id"],
-            "harvest_record_id": record.id,
-            "last_harvested_date": datetime.now(timezone.utc),
-        }
-        result = interface.upsert_dataset(reharvest_payload)
-
-        assert result.slug == user_slug
-        assert result.dcat == {"title": "Re-harvested Title"}
-
-    def test_update_dataset_slug_passes_updated_slug_to_opensearch(
-        self, interface, slug_protection_dataset
-    ):
-        """
-        Test to confirm that the updated dataset.slug is past to opensearch.
-        """
-        mock_client = MagicMock()
-        mock_client.index_datasets.return_value = (1, 0, [])
-
-        with patch(
-            "search.writer.OpenSearchWriter",
-            return_value=mock_client,
-        ):
-            interface.update_dataset_slug(
-                slug_protection_dataset.id, "reindex-slug-check"
-            )
-
-        (indexed_datasets,), _ = mock_client.index_datasets.call_args
-        assert len(indexed_datasets) == 1
-        assert indexed_datasets[0].slug == "reindex-slug-check"
-
-    def test_update_dataset_slug_logs_error_on_opensearch_failure(
-        self, interface, slug_protection_dataset, caplog
-    ):
-        """
-        When OpenSearch reports index failures, update_dataset_slug must:
-        - still return the updated dataset (DB commit succeeded)
-        - return os_synced=False
-        - return an error string describing the failure
-        - log the error at ERROR level
-        """
-        mock_client = MagicMock()
-        mock_client.index_datasets.return_value = (0, 1, [{"error": "boom"}])
-
-        with patch(
-            "search.writer.OpenSearchWriter",
-            return_value=mock_client,
-        ):
-            with caplog.at_level("ERROR"):
-                result, os_synced, os_error = interface.update_dataset_slug(
-                    slug_protection_dataset.id, "reindex-error-path"
-                )
-
-        assert result is not None
-        assert result.slug == "reindex-error-path"
-        assert os_synced is False
-        assert os_error is not None
-        assert "boom" in os_error
-        assert any("OpenSearch" in message for message in caplog.messages)
-
-    def test_update_dataset_slug_returns_error_string_on_opensearch_exception(
-        self, interface, slug_protection_dataset, caplog
-    ):
-        """
-        When OpenSearch raises an exception, the error string must contain the
-        exception message so the caller can surface it to the user.
-        """
-        with patch(
-            "search.writer.OpenSearchWriter",
-            side_effect=RuntimeError("OPENSEARCH_HOST is not set"),
-        ):
-            with caplog.at_level("ERROR"):
-                result, os_synced, os_error = interface.update_dataset_slug(
-                    slug_protection_dataset.id, "reindex-exception-path"
-                )
-
-        assert result is not None
-        assert result.slug == "reindex-exception-path"
-        assert os_synced is False
-        assert os_error is not None
-        assert "OPENSEARCH_HOST is not set" in os_error
